[tests/notebook] separate api tests into another notebook - change `mock` big query api to `simulation`
