+++ notebooks/scenarios/bigquery/apis/simulation/test_query.py
